Fix six edge-case bugs found in a full requirements re-check
- Non-interactive flag path crashed with a raw traceback on negative robot
  counts instead of the spec's clean error message.
- The default (no --level) flow raised InsufficientCapacityError before
  ever checking standby, so the spec's own Level 3 example didn't work
  through the natural entry point.
- Bonus summary printed "$43.0" instead of "$43" (float accumulator).
- A Level 4 client hitting a fully-drained pool (both active and standby
  at zero) raised NoRobotsAvailableError, which the batch loop didn't
  catch, aborting the whole run instead of skipping just that client.
- Multi-client batch exit code was always 0, even when every client
  failed, which would mislead any script checking $?.
- parse_client_hours rejected the bonus section's own bracketed notation,
  e.g. "[16, 10, 22, 7]".

Added a regression test for each. Suite: 48 -> 57 tests.

+++ robot_allocation/summary.py
@@ -7,7 +7,7 @@ from .domain import ALL_TYPES, ClientAllocation, Inventory, RobotType
 @dataclass
 class SummaryReport:
     total_robots_used: Dict[RobotType, int] = field(default_factory=dict)
-    total_charging_cost: float = 0.0
+    total_charging_cost: int = 0
     category_utilization: Dict[RobotType, float] = field(default_factory=dict)
 
     @property
@@ -24,7 +24,7 @@ class SummaryReport:
         original_standby: Inventory,
     ) -> "SummaryReport":
         used = {t: 0 for t in ALL_TYPES}
-        total_cost = 0.0
+        total_cost = 0
 
         for client in results:
             if not client.success:
